Rename Parser2 to Parser.
