+++ tests/terraform/runner/test_runner.py
@@ -1,13 +1,11 @@
 import os
 import unittest
 
-import dpath.util
-
 from checkov.runner_filter import RunnerFilter
 from checkov.terraform.context_parsers.registry import parser_registry
 from checkov.terraform.runner import Runner
 from checkov.common.output.report import Report
-from checkov.terraform.parser2 import Parser2
+from checkov.terraform.parser import Parser
 
 
 class TestRunnerValid(unittest.TestCase):
@@ -344,7 +342,7 @@ class TestRunnerValid(unittest.TestCase):
                     'current': {'start_line': 27, 'end_line': 0, 'code_lines': [], 'skipped_checks': []}}}}}
         tf_definitions = {}
         runner = Runner()
-        parser = Parser2()
+        parser = Parser()
         runner.tf_definitions = tf_definitions
         parser.parse_directory(tf_dir_path, tf_definitions)
         report = Report('terraform')
