+++ tests/terraform/test_parser2_internals.py
@@ -1,9 +1,9 @@
 import unittest
 
-from checkov.terraform import parser2
+from checkov.terraform import parser
 
 
-class TestParser2Internals(unittest.TestCase):
+class TestParserInternals(unittest.TestCase):
     def test_eval_string_to_list(self):
         expected = ["a", "b", "c"]
-        assert parser2._eval_string('["a", "b", "c"]') == expected
+        assert parser._eval_string('["a", "b", "c"]') == expected
